# backend/backend_app/static/openapi.yaml
openapi: "3.0.3"
info:
  title: API do Petshop
  description: API responsável pelo gerenciamento de usuários, pets e serviços
  version: 1.0.0
  contact: 
    name: Tiago F. Hinterholz
    url: https://github.com/tiagohinterholz/Petshop_API
    email: fh.tiago@gmail.com
servers:
  - url: http://localhost:5000
...
paths:
  /users:
    post:
      tags:
        - 'Users'
      operationId: cadastrarUser
      summary: Cadastro de Usuário
      security:
        - JWT: []
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/UsersRequest'
      responses:
        '201':
         description: "User cadastrado com sucesso"
         content:
          'application/json':
            schema:
              $ref: '#/components/schemas/UsersResponse'
        '400':
          description: "Problema nos dados de cadastro"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "Usuário não autenticado para esta operação"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse' 
        '500':
          description: "Erro inesperado ao cadastrar user"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - 'Users'
      operationId: listarUsers
      summary: Listagem de Usuários
      security:
        - JWT: []
      responses:
        '200':
          description: "Listagem de usuários realizada com sucesso"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UsersResponse'
        '401':
          description: 'Usuário não autorizado'
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'        
  /users/{usercpf}:
    get:
      tags:
        - 'Users'
      operationId: detalharUser
      summary: "Detalhar Usuário pelo CPF"
      parameters:
        - name: usercpf
          in: path
          description: "cpf do User a ser detalhado"
          required: true
          schema:
            type: string
      security:
        - JWT: []
      responses:
        '200':
          description: "User detalhado com sucesso"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersResponse"
        '400':
          description: "Dados inseridos inválidos"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "Usuario não autenticado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "Usuário não encontrado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'   
    put:
      tags:
        - 'Users'
      summary: "Atualizar cadastro de Usuário pelo CPF"
      operationId: atualizarUser
      parameters:
        - name: usercpf
          in: path
          description: "cpf do User a ser atualizado"
          required: true
          schema:
            type: string
      security:
        - JWT: []
      requestBody:
        content:
          'application/json':
            schema:
              $ref: "#/components/schemas/UsersRequest"
      responses:
        '200':
          description: "User atualizado com sucesso"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/UsersResponse'
        '400':
          description: "Dados inseridos inválidos"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "Usuario não autenticado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "Usuario não encontrado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse' 
    delete:
      tags:
        - 'Users'
      operationId: deletarUser
      summary: "Deletar Usuário pelo CPF"
      parameters:
        - name: usercpf
          in: path
          description: "Cpf do user a ser deletado"
          required: true
          schema:
            type: string
      security:
        - JWT: []
      responses:
        '200':
          description: "User deletado com sucesso"
        '401':
          description: "User não autenticado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: "User não encontrado"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "Erro ao excluir usuário"
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ErrorResponse' 
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
...
    UsersRequest:
      type: object
      required:
        - cpf
        - name
        - profile
        - password
        - email
      properties:
        cpf:
          type: string
          maxLength: 14
          example: "000.000.000-00"
        name:
          type: string
          maxLength: 100
          example: "Fulano da Silva Sauro"
        profile:
          type: string
          enum:
            - admin
            - client
          example: "admin"
        password:
          type: string
          format: password
          example: "admin1234"
        email:
          type: string
          maxLength: 120
          format: email
          example: "meuemail@email.com"
    UsersResponse:
      type: object
      properties:
        cpf:
          type: string
          example: "000.000.000-00"
        name:
          type: string
          example: "Fulano da Silva Sauro"
        profile:
          type: string
          example: "admin"
        email:
          type: string
          example: "meuemail@email.com"       
  securitySchemes:
    JWT:
      type: http
      scheme: bearer  